fix: bug fix on my-assets

+++ contracts/Market.sol
@@ -6,6 +6,7 @@ import "@openzeppelin/contracts/token/ERC721/ERC721.sol";
 import "@openzeppelin/contracts/security/ReentrancyGuard.sol";
 
 contract NFTMarket is ReentrancyGuard {
+
     using Counters for Counters.Counter;
 
     Counters.Counter private _itemIds;
@@ -52,6 +53,7 @@ contract NFTMarket is ReentrancyGuard {
         uint256 tokenId,
         uint256 price
     ) public payable nonReentrant {
+
         require(price > 0, "Price must be at least 1 wei");
         require(
             msg.value == listingPrice,
@@ -65,8 +67,8 @@ contract NFTMarket is ReentrancyGuard {
             itemId,
             nftContract,
             tokenId,
-            payable(msg.sender), // seller...
-            payable(address(0)), // owner...
+            payable(msg.sender),
+            payable(address(0)),
             price,
             false
         );
@@ -114,6 +116,7 @@ contract NFTMarket is ReentrancyGuard {
 
     /* Returns all unsold market items */
     function fetchMarketItems() public view returns (MarketItem[] memory) {
+
         uint256 itemCount = _itemIds.current();
 
         uint256 unsoldItemCount = _itemIds.current() - _itemsSold.current();
@@ -123,6 +126,7 @@ contract NFTMarket is ReentrancyGuard {
         MarketItem[] memory items = new MarketItem[](unsoldItemCount);
 
         for (uint256 i = 0; i < itemCount; i++) {
+
             if (idToMarketItem[i + 1].owner == address(0)) {
                 uint256 currentId = i + 1;
 
@@ -140,11 +144,12 @@ contract NFTMarket is ReentrancyGuard {
     /* Returns only items that a user has purchased */
     function fetchMyNFTs() public view returns (MarketItem[] memory) {
 
-        uint totalItemCount = _itemIds.current();
-        uint itemCount = 0;
-        uint currentIndex = 0;
+        uint256 totalItemCount = _itemIds.current();
+        uint256 itemCount = 0;
+        uint256 currentIndex = 0;
 
-        for (uint i = 0; i < totalItemCount; i++) {
+        for (uint256 i = 0; i < totalItemCount; i++) {
+
             if (idToMarketItem[i + 1].owner == msg.sender) {
                 itemCount += 1;
             }
@@ -152,11 +157,11 @@ contract NFTMarket is ReentrancyGuard {
 
         MarketItem[] memory items = new MarketItem[](itemCount);
 
-        for (uint i = 0; i < totalItemCount; i++) {
+        for (uint256 i = 0; i < totalItemCount; i++) {
 
             if (idToMarketItem[i + 1].owner == msg.sender) {
 
-                uint currentId = i + 1;
+                uint256 currentId = i + 1;
 
                 MarketItem storage currentItem = idToMarketItem[currentId];
 
@@ -171,6 +176,7 @@ contract NFTMarket is ReentrancyGuard {
 
     /* Returns only items a user has created */
     function fetchItemsCreated() public view returns (MarketItem[] memory) {
+
         uint256 totalItemCount = _itemIds.current();
 
         uint256 itemCount = 0;
@@ -178,6 +184,7 @@ contract NFTMarket is ReentrancyGuard {
         uint256 currentIndex = 0;
 
         for (uint256 i = 0; i < totalItemCount; i++) {
+
             if (idToMarketItem[1 + 1].seller == msg.sender) {
                 itemCount += 1;
             }
@@ -186,6 +193,7 @@ contract NFTMarket is ReentrancyGuard {
         MarketItem[] memory items = new MarketItem[](itemCount);
 
         for (uint256 i = 0; i < totalItemCount; i++) {
+
             if (idToMarketItem[i + 1].seller == msg.sender) {
                 uint256 currentId = i + 1;
